+++ docs/CHANGELOG.md
@@ -5,3 +5,4 @@ Changelog
 ----------------------
 - Init: Default driver to send code by e-mail
 - Enh: Driver "Google Authenticator"
+- Enh: Require pin code before enabling Google Authenticator